# specs/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SDD Starter",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sys/roles/{role_id}/menus": {
      "post": {
        "tags": [
          "sys_role"
        ],
        "summary": "Assign Menus",
        "operationId": "assignSysRoleMenus",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuAssign"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.2.4",
          "NFR-6.3"
        ]
      }
    },
    "/api/sys/menus": {
      "get": {
        "tags": [
          "sys_menu"
        ],
        "summary": "List Menus",
        "operationId": "listSysMenus",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuOut"
                  },
                  "title": "Response Listsysmenus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.3.1"
        ]
      },
      "post": {
        "tags": [
          "sys_menu"
        ],
        "summary": "Create Menu",
        "operationId": "createSysMenu",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.3.2",
          "FR-6.3.4"
        ]
      }
    },
    "/api/sys/menus/{menu_id}": {
      "put": {
        "tags": [
          "sys_menu"
        ],
        "summary": "Update Menu",
        "operationId": "updateSysMenu",
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Menu Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.3.2",
          "FR-6.3.4",
          "NFR-6.3"
        ]
      },
      "delete": {
        "tags": [
          "sys_menu"
        ],
        "summary": "Delete Menu",
        "operationId": "deleteSysMenu",
        "parameters": [
          {
            "name": "menu_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Menu Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.3.3",
          "NFR-6.1",
          "NFR-6.3"
        ]
      }
    },
...
    "/api/sys/logs/audit": {
      "get": {
        "tags": [
          "sys_log"
        ],
        "summary": "List Audit",
        "operationId": "listAuditLogs",
        "parameters": [
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogOut"
                  },
                  "title": "Response Listauditlogs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.4.1",
          "FR-6.4.3",
          "NFR-6.4",
          "NFR-6.6"
        ]
      }
    },
    "/api/sys/logs/login": {
      "get": {
        "tags": [
          "sys_log"
        ],
        "summary": "List Login",
        "operationId": "listLoginLogs",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Username"
            }
          },
          {
            "name": "success",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Success"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoginLogOut"
                  },
                  "title": "Response Listloginlogs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.4.2",
          "FR-6.4.3",
          "NFR-6.6"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuditLogOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "before": {
            "additionalProperties": true,
            "type": "object",
            "title": "Before"
          },
          "after": {
            "additionalProperties": true,
            "type": "object",
            "title": "After"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entity_id",
          "actor",
          "action",
          "reason",
          "before",
          "after",
          "ts"
        ],
        "title": "AuditLogOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginLogOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "ip": {
            "type": "string",
            "title": "Ip"
          },
          "user_agent": {
            "type": "string",
            "title": "User Agent"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "success",
          "ip",
          "user_agent",
          "detail",
          "ts"
        ],
        "title": "LoginLogOut"
      },
...
      "MenuAssign": {
        "properties": {
          "menu_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Menu Ids"
          }
        },
        "type": "object",
        "required": [
          "menu_ids"
        ],
        "title": "MenuAssign"
      },
      "MenuCreate": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "dir",
              "menu",
              "button"
            ],
            "title": "Type",
            "default": "menu"
          },
          "perm_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perm Code"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "order_no": {
            "type": "integer",
            "title": "Order No",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "MenuCreate"
      },
      "MenuOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "perm_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perm Code"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "order_no": {
            "type": "integer",
            "title": "Order No",
            "default": 0
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/MenuOut"
            },
            "type": "array",
            "title": "Children",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "type"
        ],
        "title": "MenuOut"
      },
      "MenuUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "dir",
                  "menu",
                  "button"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "perm_code": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Perm Code"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "order_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order No"
          }
        },
        "type": "object",
        "title": "MenuUpdate"
      },
...
      "RoleOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "remark": {
            "type": "string",
            "title": "Remark",
            "default": ""
          },
          "menu_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Menu Ids",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name"
        ],
        "title": "RoleOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}